Tables - CampaignObservation - Add date column
Adding the date column will allow us to faster query points for a
certain date than going through the PointData table. The date on the
PointData table is also a datetime, which allows for finer grained time
specifications. This removes the date column for the ImageData since
those won't need fine-grained access as the point data.