SingleLocation tables - Use a datetime throughout
Use a datetime column for the Site and PointData for consistency. This
replaces the separate time and date column for site and combines them
into one. Also adds the hybrid attribute date to the Site object to
be able to only query for dates.

Both table now carry the `datetime` column to indicate this behavior.

PR-feedback