DB Credentials - Change architecture to use environment variables.
Moving away from hardcoding a path to find the DB credentials file and use environment
variables instead. This now support two ways - one is setting the path to the file and the
other is setting the database URL directly. This should make it easier to deploy the
library and not relying on the file being in the correct location.